dqn-kuka/requirements.txt
absl-py                 2.1.0
astunparse              1.6.3
certifi                 2024.7.4
charset-normalizer      3.3.2
cloudpickle             3.0.0
contourpy               1.3.0
cycler                  0.12.1
Farama-Notifications    0.0.4
filelock                3.16.1
flatbuffers             24.3.25
fonttools               4.54.1
fsspec                  2024.9.0
gast                    0.6.0
google-pasta            0.2.0
grpcio                  1.65.2
gym                     0.26.2
gym-notices             0.0.8
gymnasium               0.29.1
h5py                    3.11.0
idna                    3.7
Jinja2                  3.1.4
keras                   3.4.1
kiwisolver              1.4.7
libclang                18.1.1
Markdown                3.6
markdown-it-py          3.0.0
MarkupSafe              2.1.5
matplotlib              3.9.2
mdurl                   0.1.2
ml-dtypes               0.4.0
mpmath                  1.3.0
namex                   0.0.8
networkx                3.3
numpy                   1.26.4
opt-einsum              3.3.0
optree                  0.12.1
packaging               24.1
pandas                  2.2.3
pillow                  10.4.0
pip                     24.2
protobuf                4.25.4
pybullet                3.2.5
Pygments                2.18.0
pyparsing               3.1.4
python-dateutil         2.9.0.post0
pytz                    2024.2
PyYAML                  6.0.2
requests                2.32.3
rich                    13.7.1
sb3_contrib             2.3.0
setuptools              75.1.0
six                     1.16.0
stable_baselines3       2.3.2
sympy                   1.13.3
tensorboard             2.17.0
tensorboard-data-server 0.7.2
tensorflow              2.17.0
tensorflow-intel        2.17.0
termcolor               2.4.0
torch                   2.4.1
typing_extensions       4.9.0
tzdata                  2024.2
urllib3                 1.26.13
Werkzeug                2.0.3
wheel                   0.43.0
wrapt                   1.16.0
yacs                    0.1.8
zipp                    3.6.0
